core: adapt to datascript Int64 values (pin e439cba)
Datascript upstream c39cad1 changed value Int of int to Int64 of int64.
Convert at the boundary: chat domain types stay int, wrapped via
Int64.of_int/Int64.to_int; test literals get the L suffix; sync codec
emits Transit.Int64 unconditionally like upstream. file/created-at and
file/last-modified-at are written as Int64 rather than Instant since
upstream now demotes Instant to Int64 for attrs without
:db.type/instant on storage read.

+++ shared/test/logseq_chat/graph_store_test.ml
@@ -130,7 +130,7 @@ let typed_restore_and_writable_connection_persist () =
       (values db "block/uuid");
     check_eq [ Ds.Ref 11 ] (values db "block/parent");
     check_eq [ Ds.Keyword "whiteboard" ] (values db "block/type");
-    check_eq [ Ds.Int 1723012345678 ] (values db "block/created-at");
+    check_eq [ Ds.Int64 1723012345678L ] (values db "block/created-at");
     check_eq [ Ds.Map [ (Ds.Keyword "priority", Ds.Keyword "A") ] ]
       (values db "block/properties");
     check_eq [ Ds.Ref 11; Ds.Ref 12 ] (values db "block/tags");
